First cut support for init script command line argument support.
Introduces a backwards compatible command line syntax enhancement
that supports passing arguments to init script commands. If an
init script command (e.g., stop, start, restart) has a trailing
colon (':'), then the following argv values are treated as
arguments up to (but not including) the next argv value with a
trailing colon (or the end of arguments, whichever comes first).
Example:

	service foo start: bar baz

Service foo is started with arguments "bar baz" which are passed
in a new environment variable RC_ARGS.

It's important to emphasize that existing init script invocation
syntax has not been disturned in any way by this enhanced syntax
implementation.

This is merely a first cut incomplete proof of concept. Still
to do:

- Alternate syntax proposal implementations.
- Usage help text.
- Code to flag init scripts that do not take arguments. Arguments
  are harmless to init scripts that do not take them, but I'd like
  to report when arguments are being ignored.