Work on alternate command line syntax using -a.
Working on a variation of the argument support command line
syntax that loses the multiple commands per line feature but more
simply supports arguments:

        service -a foo start bar baz

Service foo is started with arguments "bar baz" which, as before,
are passed in a new environment variable RC_ARGS. Unlike the
other proposals, there's no way to specify multiple service
commands in a single invocation.

Not complete; work in progress.