rc-service: More usage text reorganization / code style cleanup.
The getopts options string is used directly in the usage text.
Although the order that the options appear in that string does
not affect getopt()'s ability to parse command line options, it
looks funny to have it presented in arbitrary order (most likely
the order in which the options were added to the code) in the
usage text, so I cleaned it up.